tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animationevent_inheritance.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - AnimationEvent interface inheritance</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-events-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The AnimationEvent interface inherits from
   Event">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = false;
         try {
            var evt = document.createEvent('AnimationEvent');
            passed = (evt.CAPTURING_PHASE == 1);
         } catch(e) {
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = passed ? 'PASS' : 'FAIL';
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
</body>
</html>
